Refactored stripe api into modules
