+++ packages/members-api/lib/subscriptions/payment-processors/stripe/api/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+    customers: require('./customers'),
+    products: require('./products'),
+    plans: require('./plans'),
+    subscriptions: require('./subscriptions')
+};
