+++ packages/members-api/lib/subscriptions/payment-processors/stripe/api/products.js
@@ -1,0 +1,12 @@
+const createDeterministicApi = require('./createDeterministicApi');
+
+const isActive = x => x.active;
+const getProductAttr = ({name}) => ({name, type: 'service'});
+const getProductHashSeed = () => 'Ghost Subscription';
+
+module.exports = createDeterministicApi(
+    'products',
+    isActive,
+    getProductAttr,
+    getProductHashSeed
+);
